pyproject.toml
[project]
name = "valutatrade_hub"
version = "0.1.0"
description = ""
authors = [
    {name = "Anastasiya Osadchuk",email = "diamondtheb@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prettytable (>=3.16.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

packages = [
    {include = "valutatrade_hub"},
    ]

[tool.poetry.scripts]
project = "main:main"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["venv", ".venv", "dist", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.2,<0.15.0)"
]
